SCB-1368 Refactoring model alpha-fsm-channel-kafka and alpha-fsm-channel-redis to alpha-fsm
