+++ alpha/pom.xml
@@ -33,8 +33,6 @@
   <modules>
     <module>alpha-core</module>
     <module>alpha-fsm</module>
-    <module>alpha-fsm-channel-redis</module>
-    <module>alpha-fsm-channel-kafka</module>
     <module>alpha-benchmark</module>
     <module>alpha-spring-cloud-starter-eureka</module>
     <module>alpha-spring-cloud-starter-consul</module>
